dietapp/requirements.txt
Django>=5.2
djangorestframework
djangorestframework-simplejwt
django-cors-headers
django-crontab

